Move CopySchemaUtil and its test to util package and make CopySchemaUtil public
